Added: failure.py
